Remove list action.
After ensuring that any valid paths could be handled, listing the
snapshots became too problematic. There's no good way to list files that
could potentially have newlines (even if it's a remote possibility).
Since the list action is not crucial to the way the application
function, I decided to remove it.